style: rename bahaviours
