+++ test/examples/ERC1363MethodCallReceiver.behavior.js
